# apps/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Zone Management System API",
    "description": "REST API for managing a gaming cafe: users, devices, plans, sessions, transactions, and inventory.",
    "contact": {
      "name": "Gaming Cafe Team"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/files": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "list_files",
        "responses": {
          "200": {
            "description": "List files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePaginationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "files"
        ],
        "operationId": "create_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/files/stats": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get_storage_stats",
        "responses": {
          "200": {
            "description": "Get file storage statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageStatsEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/files/{id}": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "files"
        ],
        "operationId": "update_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "files"
        ],
        "operationId": "delete_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Soft-deactivated (status=deleted)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/files/{id}/activate": {
      "put": {
        "tags": [
          "files"
        ],
        "operationId": "activate_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activate file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/files/{id}/archive": {
      "put": {
        "tags": [
          "files"
        ],
        "operationId": "archive_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/files/{id}/download-url": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get_file_download_url",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expires_in",
            "in": "query",
            "description": "URL expiry in seconds",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get file download URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileWithDownloadUrlEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFileDto": {
        "type": "object",
        "required": [
          "fileName",
          "originalFileName",
          "storageKey",
          "contentType",
          "fileSize"
        ],
        "properties": {
          "bucket": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "contentType": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "extension": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "metadata": {},
          "originalFileName": {
            "type": "string"
          },
          "relatedEntityId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "relatedEntityType": {
            "type": [
              "string",
              "null"
            ]
          },
          "storageKey": {
            "type": "string"
          },
          "tags": {},
          "visibility": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
      "FileEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FileRecord"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "FilePaginationEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FilePaginationPage"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
      "FilePaginationPage": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRecord"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FileWithDownloadUrlDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FileRecord"
          },
          {
            "type": "object",
            "required": [
              "downloadUrl",
              "urlExpiresIn"
            ],
            "properties": {
              "downloadUrl": {
                "type": "string"
              },
              "urlExpiresIn": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          }
        ]
      },
      "FileWithDownloadUrlEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FileWithDownloadUrlDto"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "StorageStatsDto": {
        "type": "object",
        "required": [
          "totalFiles",
          "totalSize",
          "byCategory",
          "byStatus"
        ],
        "properties": {
          "byCategory": {},
          "byStatus": {},
          "totalFiles": {
            "type": "integer",
            "format": "int64"
          },
          "totalSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StorageStatsEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StorageStatsDto"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "UpdateFileDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileName": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {},
          "relatedEntityId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "relatedEntityType": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {},
          "visibility": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /auth/verify-otp"
      }
    }
  },
...
}